+++ code/tempTrender.cpp
@@ -13,6 +13,7 @@
 #define NB_MAX_FILE_VALUES 200000
 #define MAX_YEAR 2100 // maximum year of input data
 #define MIN_TEMP_VALUE -100 // minimum value of temperature
+#define ARRAY_SIZE 100000 // minimum value of temperature
 
 /*
  * Constructor
@@ -25,6 +26,7 @@ tempTrender::tempTrender(std::string filePath) {
         //cout << "You should probably store this information in a member variable of the class. Good luck with the project! :)" << endl;
 }
 
+
 //////////////////////////////////
 // JASMAIN CODE PART
 //////////////////////////////////
@@ -35,6 +37,94 @@ tempTrender::tempTrender(std::string filePath) {
  *               (year, month, day, temp, temp with urban correction, city)
  * @return map of mean temp per year
  */
+std::map<int, double> coldTempPerYear(std::vector<std::string> entries) {
+        // temporary values
+        std::string year_str, value_str;
+        size_t pos;
+        int year;
+        double value;
+        
+        // storage
+        double cold_values[MAX_YEAR]={0};
+        
+        // for each entry, extract year and value
+        for(size_t y=0; y<entries.size(); y++) {
+                std::string s = entries[y];
+                // get year from 4 first characters
+                year_str = s.substr(0, 4);
+                year = atoi(year_str.c_str());                             
+                
+                // get value from 5th field
+                for(size_t i=0; i<4; i++) {
+                        pos = s.find(' '); // find first delimiter
+                        s.substr(0, pos);
+                        s.erase(0, pos + 1);
+                }
+                pos = s.find(' ');
+                value_str = s.substr(0, pos);
+                value = atof(value_str.c_str());
+                
+                // add value to arrays
+                if (value < cold_values[year] ) {
+                    cold_values[year] = value;
+                }
+        }
+        
+        // return variable
+        std::map<int, double> ret;
+        double cold;
+        
+        // for each year, compute mean value
+        for(size_t i=0; i<MAX_YEAR; i++) {
+               cold = cold_values[i];
+               std::cerr << "year : " << i << " coldest temp : " << cold << std::endl;
+               ret.insert(std::make_pair<int, double>((int)i, (double)cold));
+        }
+        return ret;
+}
+void tempTrender::coldestTempPerYear() {
+        
+        // open input file
+        std::string line;
+        ifstream inputfile(_filePath.c_str());
+        
+        
+        // create array to store values
+        std::vector<std::string> entries;
+        
+        
+        // read values and stor it into array, hist or something else
+        if(inputfile.is_open()) {
+                while(getline(inputfile, line)) {
+                        entries.push_back(line);
+                }
+        }
+
+        std::map<int, double> cold = coldTempPerYear(entries);
+        
+        // create canvas for graph
+        TCanvas *c2 = new TCanvas("Jasmain", "Project : COldest temperature over each year");
+
+        // create new histogram object
+        TH1D* hist2 = new TH1D("hist", "Coldest Temp Per Year", cold.size(), 1722, 2013);
+        
+        for( std::map<int, double>::iterator it = cold.begin();
+             it != cold.end();
+             it++ ) {
+                // fill hist with date and value from mean temp per year map
+                //std::cerr << "value : " << it->second << std::endl;
+                hist2->Fill(it->first, it->second);
+        }
+        //Axis title
+        hist2->SetTitle("Coldest temperature over each year (Uppsala, 1722-2013)");
+        hist2->GetXaxis()->SetTitle("Year");
+        hist2->GetXaxis()->CenterTitle();
+        hist2->GetYaxis()->SetTitle("Coldest Temperature (Celsius deg)");
+        hist2->GetYaxis()->CenterTitle();
+
+        // draw hist
+        hist2->Draw("SAME");
+}
 std::map<int, double> hotTempPerYear(std::vector<std::string> entries) {
         // temporary values
         std::string year_str, value_str;
@@ -45,17 +135,12 @@ std::map<int, double> hotTempPerYear(std::vector<std::string> entries) {
         // storage
         double hot_values[MAX_YEAR]={0};
         
-        // initialise hot_values to have a minimum value
-        for(size_t i=0; i<MAX_YEAR; i++) {
-            hot_values[i] = MIN_TEMP_VALUE;
-        }
-        
         // for each entry, extract year and value
         for(size_t y=0; y<entries.size(); y++) {
                 std::string s = entries[y];
                 // get year from 4 first characters
                 year_str = s.substr(0, 4);
-                year = atoi(year_str.c_str());
+                year = atoi(year_str.c_str());                             
                 
                 // get value from 5th field
                 for(size_t i=0; i<4; i++) {
@@ -80,12 +165,12 @@ std::map<int, double> hotTempPerYear(std::vector<std::string> entries) {
         // for each year, compute mean value
         for(size_t i=0; i<MAX_YEAR; i++) {
                hot = hot_values[i];
-               std::cerr << "year : " << i << " hottest temp : " << hot << std::endl;
+               //std::cerr << "year : " << i << " hottest temp : " << hot << std::endl;
                ret.insert(std::make_pair<int, double>((int)i, (double)hot));
         }
         return ret;
 }
-void tempTrender::hottestTempPerYear(int yearToExtrapol) {
+void tempTrender::hottestTempPerYear() {
         
         // open input file
         std::string line;
@@ -109,10 +194,10 @@ void tempTrender::hottestTempPerYear(int yearToExtrapol) {
         TCanvas *c1 = new TCanvas("Jasmain", "Project : Hottest temperature over each year");
 
         // create new histogram object
-        TH1D* hist = new TH1D("hist", "Hottest Temp Per Year", hotPerYear.size(), 1722, 2013);
+        TH1D* hist = new TH1D("hist", "Hottest Temp Per Year", hot.size(), 1722, 2013);
         
-        for( std::map<int, double>::iterator it = hotPerYear.begin();
-             it != hotPerYear.end();
+        for( std::map<int, double>::iterator it = hot.begin();
+             it != hot.end();
              it++ ) {
                 // fill hist with date and value from mean temp per year map
                 //std::cerr << "value : " << it->second << std::endl;
@@ -136,22 +221,18 @@ void tempTrender::hottestTempPerYear(int yearToExtrapol) {
 // JOHAN CODE PART
 //////////////////////////////////
 
-// NOTE : you should give more explicit names to your variables
-// Single letter variables are quite hard to maintain, prefer longer names
-
-// Be careful at which file is used
 
 void tempTrender::tempOnDay(int monthToCalculate, int dayToCalculate) {
-    
+
     ifstream file(_filePath.c_str());
     
-    //Defining all vaiables we need
+    //Defining all variables we need
     vector<string> entries;
-    vector<float> temp;
+    float temp[ARRAY_SIZE];
     
     string line, month_str, day_str, temp_str;
     int i, j=0, size, pos, month, day;
-    float t;
+    float t, mean=0, StanDev=0;
     
     //Check if the file is opened
     if(file.is_open()) 
@@ -168,7 +249,7 @@ void tempTrender::tempOnDay(int monthToCalculate, int dayToCalculate) {
 
     
     
-    TCanvas *c1= new TCanvas("Johan","Projet : ");
+    TCanvas *c1= new TCanvas("Johan","Project : Temperature on a given day");
     
     //creating histogram
     TH1D* hist = new TH1D("Hist", "Temperature on a given day", 100, -20, 40);
@@ -201,12 +282,38 @@ void tempTrender::tempOnDay(int monthToCalculate, int dayToCalculate) {
             pos = line.find(' ');
             temp_str = line.substr(0, pos);
             t = atof(temp_str.c_str());
+            mean=mean+t;
+            temp[j]=t;
+            j=j+1;
             hist->Fill(t);
         }
     }
+    
+    
+    //Calculate the mean value on that day
+    mean=mean/j;
+    
+    //Calculate the standart deviation
+    for (i=0; i<=j; i++){
+        StanDev = StanDev+(temp[i]-mean)*(temp[i]-mean)/(j+1);
+    }
+    StanDev = sqrt(StanDev);
+    
+    cout << "The mean value on this day is : " << mean << endl;
+    cout << "The standart deviation is : " << StanDev << endl;
+    
+    //Setting up X and Y axis
+    hist->GetYaxis()->SetTitle("Entries");
+    hist->GetXaxis()->SetTitle("Temperature in Celsius (deg)");
+        
+
     //drawing histogram
+    hist->SetFillColor(kGreen);
     hist->Draw("SAME");
+    
+
 }
+
 
 
 
@@ -282,7 +389,7 @@ void tempTrender::tempPerYear(int yearToExtrapolate) {
         // create array to store values
         std::vector<std::string> entries;
         
-        // read values and store it into array, hist or something else
+        // read values and store into entries array
         if(inputfile.is_open()) {
                 while(getline(inputfile, line)) {
                         entries.push_back(line);
@@ -332,8 +439,7 @@ void tempTrender::tempPerYear(int yearToExtrapolate) {
             graph->SetPoint(graph->GetN(), hist->GetBinCenter(bin),
                 hist->GetBinContent(bin));
         }
-        //graph->Draw("SAME C");
-        
+
         // create function for extrapolation
         TF1 *f = (TF1*)hist->GetFunction("pol7");
         //f->Eval(yearToExtrapolate);