+++ code/tempTrender.h
@@ -12,7 +12,8 @@ public:
     //void tempPerDay(); //Make a histogram of the average temperature of each day of the year
     //void hotCold(); //Make a histogram of the hottest and coldest day of the year
     void tempPerYear(int yearToExtrapolate); //Make a histogram of average temperature per year, then fit and extrapolate to the given year
-    //void hottestTempPerYear(int yearToExtrapol); //Make a histogram of hottest temperature per year, then fit (maybe if possible extrapolate to the given year)
+    void hottestTempPerYear(); //Make a histogram of hottest temperature per year, then fit (maybe if possible extrapolate to the given year)
+    void coldestTempPerYear();
 private:
     std::string _filePath;